code cleanup

+++ src/main/java/at/hannibal2/skyhanni/features/event/diana/GriffinBurrowHelper.kt
@@ -191,7 +191,7 @@ object GriffinBurrowHelper {
                 event.drawColor(location, LorenzColor.LIGHT_PURPLE)
                 val distance = location.distance(playerLocation)
                 if (distance > 10) {
-                    val formattedDistance = distance.toInt().addSeparators()
+                    val formattedDistance = distance.addSeparators()
                     event.drawDynamicText(location.add(y = 1), "§d§lInquisitor §e${formattedDistance}m", 1.7)
                 } else {
                     event.drawDynamicText(location.add(y = 1), "§d§lInquisitor", 1.7)
@@ -253,7 +253,7 @@ object GriffinBurrowHelper {
                 val color = if (currentWarp == null) "§f" else "§b"
                 event.drawDynamicText(guessLocation.add(y = 1), "${color}Guess", 1.5)
                 if (distance > 5) {
-                    val formattedDistance = distance.toInt().addSeparators()
+                    val formattedDistance = distance.addSeparators()
                     event.drawDynamicText(guessLocation.add(y = 1), "§e${formattedDistance}m", 1.7, yOff = 10f)
                 }
             }
